# openapi-schema.json
openapi: 3.0.3
info:
  title: ChargeGhar Nepal API
  version: a315307-dirty
  description: ChargeGhar API - Shared Power Bank Network for Nepal. This documentation
    only shows endpoints documented in our Features TOC.
  contact:
    name: ChargeGhar API Support
    email: support@chargegh.com
  license:
    name: Proprietary
paths:
...
  /api/auth/account:
    delete:
      operationId: api_auth_account_destroy
      description: Permanently deletes user account and data
      summary: Delete Account
      tags:
      - Authentication
      security:
      - cookieAuth: []
      - basicAuth: []
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseResponse'
          description: ''
...
  /api/auth/device:
    post:
      operationId: api_auth_device_create
      description: Update FCM token and device data
      summary: Register Device
      tags:
      - Authentication
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/UserDeviceRequest'
          application/json:
            schema:
              $ref: '#/components/schemas/UserDeviceRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/UserDeviceRequest'
        required: true
      security:
      - cookieAuth: []
      - basicAuth: []
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseResponse'
          description: ''
...
  /api/auth/login:
    post:
      operationId: api_auth_login_create
      description: Completes login after OTP verification (no password required)
      summary: User Login
      tags:
      - Authentication
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/UserLoginRequest'
          application/json:
            schema:
              $ref: '#/components/schemas/UserLoginRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/UserLoginRequest'
        required: true
      security:
      - cookieAuth: []
      - basicAuth: []
      - jwtAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseResponse'
          description: ''
  /api/auth/logout:
    post:
      operationId: api_auth_logout_create
      description: Invalidates JWT and clears session
      summary: User Logout
      tags:
      - Authentication
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/UserLoginRequest'
          application/json:
            schema:
              $ref: '#/components/schemas/UserLoginRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/UserLoginRequest'
        required: true
      security:
      - cookieAuth: []
      - basicAuth: []
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseResponse'
          description: ''
  /api/auth/me:
    get:
      operationId: api_auth_me_retrieve
      description: Returns authenticated user's real-time data
      summary: Current User Info
      tags:
      - Authentication
      security:
      - cookieAuth: []
      - basicAuth: []
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseResponse'
          description: ''
  /api/auth/otp/request:
    post:
      operationId: api_auth_otp_request_create
      description: Sends OTP via SMS or Email for registration or login
      summary: Request OTP
      tags:
      - Authentication
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/OTPRequestRequest'
          application/json:
            schema:
              $ref: '#/components/schemas/OTPRequestRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/OTPRequestRequest'
        required: true
      security:
      - cookieAuth: []
      - basicAuth: []
      - jwtAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseResponse'
          description: ''
  /api/auth/otp/verify:
    post:
      operationId: api_auth_otp_verify_create
      description: Validates OTP and returns verification token for registration or
        login
      summary: Verify OTP
      tags:
      - Authentication
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/OTPVerificationRequest'
          application/json:
            schema:
              $ref: '#/components/schemas/OTPVerificationRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/OTPVerificationRequest'
        required: true
      security:
      - cookieAuth: []
      - basicAuth: []
      - jwtAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseResponse'
          description: ''
  /api/auth/refresh:
    post:
      operationId: api_auth_refresh_create
      description: Refreshes JWT access token
      summary: Refresh Token
      tags:
      - Authentication
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/TokenRefreshRequest'
          application/json:
            schema:
              $ref: '#/components/schemas/TokenRefreshRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/TokenRefreshRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenRefresh'
          description: ''
  /api/auth/register:
    post:
      operationId: api_auth_register_create
      description: Creates new user account after OTP verification (no password required)
      summary: User Registration
      tags:
      - Authentication
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/UserRegistrationRequest'
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegistrationRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/UserRegistrationRequest'
        required: true
      security:
      - cookieAuth: []
      - basicAuth: []
      - jwtAuth: []
      - {}
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseResponse'
          description: ''
...
components:
  schemas:
...
    BaseResponse:
      type: object
      description: Standard response format for all API endpoints
      properties:
        success:
          type: boolean
        message:
          type: string
        data: {}
        timestamp:
          type: string
          format: date-time
      required:
      - data
      - message
      - success
      - timestamp
...
    OTPRequestRequest:
      type: object
      description: Serializer for OTP request
      properties:
        identifier:
          type: string
          minLength: 1
          description: Email or phone number
        purpose:
          enum:
          - REGISTER
          - LOGIN
          type: string
          description: |-
            * `REGISTER` - REGISTER
            * `LOGIN` - LOGIN
          x-spec-enum-id: 5460c80c127fa7b3
      required:
      - identifier
      - purpose
    OTPVerificationRequest:
      type: object
      description: Serializer for OTP verification
      properties:
        identifier:
          type: string
          minLength: 1
          description: Email or phone number
        otp:
          type: string
          minLength: 6
          maxLength: 6
        purpose:
          enum:
          - REGISTER
          - LOGIN
          type: string
          description: |-
            * `REGISTER` - REGISTER
            * `LOGIN` - LOGIN
          x-spec-enum-id: 5460c80c127fa7b3
      required:
      - identifier
      - otp
      - purpose
...
    TokenRefresh:
      type: object
      properties:
        access:
          type: string
          readOnly: true
        refresh:
          type: string
      required:
      - access
      - refresh
    TokenRefreshRequest:
      type: object
      properties:
        refresh:
          type: string
          minLength: 1
      required:
      - refresh
...
    UserDeviceRequest:
      type: object
      description: Serializer for user device registration
      properties:
        device_id:
          type: string
          minLength: 1
          maxLength: 255
        fcm_token:
          type: string
          minLength: 1
        device_type:
          enum:
          - ANDROID
          - IOS
          - WEB
          type: string
          description: |-
            * `ANDROID` - Android
            * `IOS` - iOS
            * `WEB` - Web
          x-spec-enum-id: 5617f81b72d11de6
        device_name:
          type: string
          nullable: true
          maxLength: 255
        app_version:
          type: string
          nullable: true
          maxLength: 50
        os_version:
          type: string
          nullable: true
          maxLength: 50
        is_active:
          type: boolean
      required:
      - device_id
      - device_type
      - fcm_token
...
    UserLoginRequest:
      type: object
      description: Serializer for OTP-based user login
      properties:
        identifier:
          type: string
          minLength: 1
          description: Email or phone number
        verification_token:
          type: string
          writeOnly: true
          minLength: 1
      required:
      - identifier
      - verification_token
...
    UserRegistrationRequest:
      type: object
      description: Serializer for OTP-based user registration
      properties:
        identifier:
          type: string
          minLength: 1
          description: Email or phone number
        username:
          type: string
          maxLength: 150
        referral_code:
          type: string
          maxLength: 10
        verification_token:
          type: string
          writeOnly: true
          minLength: 1
      required:
      - identifier
      - verification_token
...
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...